+++ src/cli.rs
@@ -30,19 +30,19 @@ pub struct CommonArgs {
     #[arg(long, overrides_with("drop_empty_cells"), hide = true)]
     pub keep_empty_cells: bool,
 
-    #[arg(long, overrides_with("keep_output"), hide = true)]
-    pub drop_output: bool,
-
     /// keep cell output. Disable with `--drop-output`
     #[arg(long, overrides_with("drop_output"))]
     pub keep_output: bool,
 
-    #[arg(long, overrides_with("keep_count"), hide = true)]
-    pub drop_count: bool,
+    #[arg(long, overrides_with("keep_output"), hide = true)]
+    pub drop_output: bool,
 
     /// keep cell exeution count. Disable with `--drop count`
     #[arg(long, overrides_with("drop_count"))]
     pub keep_count: bool,
+
+    #[arg(long, overrides_with("keep_count"), hide = true)]
+    pub drop_count: bool,
 
     /// replace cell ids with sequential ids. Disable with `--keep-id`
     #[arg(long, overrides_with("keep_id"))]
@@ -77,6 +77,8 @@ pub enum Commands {
     Check(CheckCommand),
     /// clean a single notebook
     Clean(CleanCommand),
+    /// uninstall nbwipers as a git filter
+    Uninstall(UninstallCommand),
 }
 
 #[derive(Clone, Debug, Parser)]
@@ -117,6 +119,15 @@ pub struct CleanAllCommand {
 }
 #[derive(Clone, Debug, Parser)]
 pub struct InstallCommand {
+    /// Git config type that determines which file to modify
+    #[clap(value_enum)]
+    pub config_type: GitConfigType,
+    /// optional attribute file. If not specified, will write to .git/info/attributes
+    #[arg(long, short)]
+    pub attribute_file: Option<PathBuf>,
+}
+#[derive(Clone, Debug, Parser)]
+pub struct UninstallCommand {
     /// Git config type that determines which file to modify
     #[clap(value_enum)]
     pub config_type: GitConfigType,